app: ui: improvements
